requirements.txt
tensorflow-gpu==2.4.1
opencv-python==4.5.1.48
scikit-learn==0.24.1
numpy==1.19.5
imgaug==0.4.0
matplotlib==3.4.1